default /healthcheck based on ServiceManager
Previously the `/healthcheck` endpoint provided here always
returned 200 and was mostly used for debugging. This is now
`/ok`. The `/healthcheck` endpoint now indicates the service
health based on the ServiceManager so it should be usable
as a load balancer healthcheck endpoint.

With this change we shouldn't need a separate internal
healthcheck endpoint registered for use with the load
balancers and monitoring.