WIP: experiment for suppoorting box and whisker
For canary reports emails there is some desire to have a
simple rendering of box and whisker plots for outliers.
This can be done with some hacks on top of jfreechart, but
is a bit hacky and hard to get right. It isn't really
designed for producing an image directly.

This is just a quick experiment to try rendering them using
some of the Atlas charting components.