parse messages after merge of streams
Moves the step for parsing messages after the merge of
streams from each of the LWCAPI nodes. This makes it
easier to keep track of the overall number of concurrent
requests and ensure it backpressures rather than enqueues.

Also changes to use the ordered variant. Preserving the
order could be important to avoid newer messages getting
pushed through first and moving the cutoff point for the
time grouping.